separate sign in component and rename signin route to authentication
