+++ src/App.js
@@ -3,7 +3,7 @@ import { Routes, Route } from "react-router-dom";
 import Home from "./routes/home/Home";
 import Navigation from "./routes/navigation/Navigation";
 import Shop from "./routes/shop/Shop";
-import SignIn from "./routes/signin/SignIn";
+import Authentication from "./routes/authentication/Authentication";
 
 
 const App = () => {
@@ -12,7 +12,7 @@ const App = () => {
       <Route path="/" element={<Navigation />}>
         <Route index element={<Home />} />
         <Route path="/" element={<Home />} />
-        <Route path="sign-in" element={<SignIn />} />
+        <Route path="auth" element={<Authentication />} />
         <Route path="shop" element={<Shop />} />
       </Route>
     </Routes>
