+++ src/routes/authentication/Authentication.jsx
@@ -1,0 +1,13 @@
+import Signin from "../../components/sign-in-form/Signin";
+import SignUp from "../../components/sign-up-form/SignUp";
+
+const Authentication = () => {
+  return (
+    <>
+      <Signin />
+      <SignUp />
+    </>
+  );
+};
+
+export default Authentication;
