+++ src/routes/navigation/Navigation.jsx
@@ -17,7 +17,7 @@ const Navigation = () => {
           <Link className="nav-link" to="/shop">
             Shop
           </Link>
-          <Link className="nav-link" to="/sign-in">
+          <Link className="nav-link" to="/auth">
             Sign in
           </Link>
         </div>
